requirements.txt
# ---- Base ----
django
djangorestframework
fastapi
uvicorn[standard]

# ---- DB ----
psycopg2-binary # PostgreSQL Driver

# ---- AI Stack ----
langchain
langchain-community
langchain-core
ollama

# ---- Input Processing ----
docling
langchain-docling